Add test

+++ src/main/java/com/ntnghia/todo/controllers/TodoController.java
@@ -1,5 +1,6 @@
 package com.ntnghia.todo.controllers;
 
+import com.ntnghia.todo.exceptions.NotFoundException;
 import com.ntnghia.todo.models.Todo;
 import com.ntnghia.todo.repositories.TodoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,17 +17,23 @@ public class TodoController {
     private TodoRepository todoRepository;
 
     @GetMapping()
-    List<Todo> getAll(){
+    List<Todo> getAll() {
         return todoRepository.findAll();
     }
 
     @GetMapping("/{id}")
-    Optional<Todo> get(@PathVariable int id){
-        return todoRepository.findById(id);
+    Optional<Todo> get(@PathVariable int id) {
+        Optional<Todo> optionalTodo = todoRepository.findById(id);
+
+        if (optionalTodo.isPresent()) {
+            return optionalTodo;
+        }
+
+        throw new NotFoundException(String.format("Todo id %d not found", id));
     }
 
     @PostMapping()
-    Todo post(@RequestBody Todo todo){
+    Todo post(@RequestBody Todo todo) {
         todo.setId(0);
         todoRepository.save(todo);
 
@@ -34,12 +41,16 @@ public class TodoController {
     }
 
     @PutMapping()
-    void put(@RequestBody Todo todo){
+    void put(@RequestBody Todo todo) {
         todoRepository.save(todo);
     }
 
     @DeleteMapping("/{id}")
-    void delete(@PathVariable int id){
+    void delete(@PathVariable int id) {
+        if (!todoRepository.existsById(id)) {
+            throw new NotFoundException(String.format("Book id %d not found", id));
+        }
+
         todoRepository.deleteById(id);
     }
 }
